* Fixed reading of aromatic bond from MDL molfiles   (fixes bug #620731, thanks to Hans Vanderzijl for patch   suggestion) * Changed order parameter of two AtomContainer.addBond() methods   from int to double signature (fixes bug #620241)
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@765 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/AtomContainer.java
@@ -898,7 +898,7 @@ public class AtomContainer extends ChemObject implements Cloneable
 	 * @param  order   Bondorder
 	 * @param  stereo  Stereochemical orientation
 	 */
-	public void addBond(int atom1, int atom2, int order, int stereo)
+	public void addBond(int atom1, int atom2, double order, int stereo)
 	{
 		Bond bond = new Bond(getAtomAt(atom1), getAtomAt(atom2), order, stereo);
 
@@ -922,7 +922,7 @@ public class AtomContainer extends ChemObject implements Cloneable
 	 * @param  atom2  Id of the second atom of the Bond
 	 * @param  order  Bondorder
 	 */
-	public void addBond(int atom1, int atom2, int order)
+	public void addBond(int atom1, int atom2, double order)
 	{
 		Bond bond = new Bond(getAtomAt(atom1), getAtomAt(atom2), order);
